Make rocks more common, and clean up after the Wiz of Yendor change.
There are now more rocks in the dungeon, courtesy of mineralize().
The precise formula will probably need to be adjusted, but there is
now a good framework in place for placing them.

This is intended to make sling a less horrible option, compared to
other ranged weapons.  It's still inferior, because rocks aren't
particularly good ammo -- flint is the good ammo for slings.  But
hopefully the player should now at least not completely run out of
things to sling.

I also cleaned up some variable name issues in mcastu.c, left over
from the addition of the Wizard of Yendor's new attack that changes
your alignment.