Apply Dev Team fix, C343-257, nethack.rankin, Sat Feb 10 05:14:22 2007 +0000
Several places in the code treated any cursed object that a monster
was wielding as welded.  One way to reproduce the bug was to apply a
bullwhip toward a monster that was wielding a cursed cockatrice
corpse.  The fix adds a routine, mwelded, to check whether the item is
in fact welded or not; corpses are not.

In order to avoid applying unrelated changes to how monster wear/wield
slots are tracked (or possibly de-applying them; I'm not sure which
codebase diverged from the other here), I made the NH4 version of
mwelded() take the monster and the object as arguments, rather than
just the object.  This hopefully also obviates the risk that someone
will call the routine on a random object without checking that it
belongs to a monster at all.

There was an additional bit in the patch that I did not apply at this
time, because it applies to an earlier fix (the munstone commit from
Thu Apr 13 07:10:48 2006 +0000).  If we apply that fix to NH4 later,
mcould_eat_tin should be adjusted to use mwelded then.