Ticket #436, C343-169: apply Dev Team fix, nethack.rankin, Thu Apr 6 04:55:07 2006 +0000.
A player had reported that moving onto a trap as a result of a
failed #untrap attempt, when there were already objects on the
trap tile that autopickup would normally pick up, did not result
in any objects being picked up or any message about the objects.

The person who wrote that bit of code did not remember whether
the lack of pickup was deliberate; perhaps it was to avoid a
redundant message, but the fix avoids that by setting tseen = 0;
apart from that, passing TRUE rather than FALSE to spoteffects()
causes it to go ahead and pick things up.