Apply Dev Team fix, UNL-046, nethack.rankin, Tue Feb 20 02:32:50 2007 +0000
You can't lock or unlock a chest on the floor while engulfed, but
you could lock or unlock an adjacent door.  The fix gives you a
special error message telling you that you can't lock or unlock
the monster that is engulfing you.

Applied straightforwardly.  The only changes to the patch for NH4 are
using Engulfed (instead of u.uswallow directly) and not having
wrapping the credit card check (to exclude "lock or ") with #ifdef
TOURIST, due to the great NitroHack #define purge.