Ticket 588, C343-262, apply first part of Dev Team fix, cohrs, Mon Feb 12 21:25:52 2007 +0000
If the player polymorphs into something that cannot wield a weapon,
such as a killer bee, the game forces you to drop your weapon.
Normally that's fine, but for some reason the game allows you to wield
an in-use leash; if the player was doing so when the polymorph occurs,
this resulted in dropping an in-use leash without removing its in-use
status or releasing the pet from it.

The fix unwields the leash but leave it in inventory and attached to
you, as would happen if the leash were not wielded.

I went ahead and corrected the logic on the test.  (The original
author corrected it as part of the next change, but I didn't see
any point in committing an incorrect test.)

This commit does not handle the two-weapon case correctly; that is
coming in a second commit.

It is still possible to wield an in-use leash.