Ticket 588, C343-262, apply second part of Dev Team fix, cohrs, Mon Feb 12 21:40:16 2007 +0000
This is the second, two-weapon portion of the fix for dropping a
wielded in-use leash due to polymorphing into a non-wielding form.
I'm not sure this can actually happen, because normally you can't
engage in two-weapon combat unless both weapons are in fact weapons.
But I applied the fix anyway; if nothing else, it keeps the code
for the two cases similar.

The original patch also contained a correction to the previous patch
(handling the primary weapon case), but I already rolled that
correction into my application of that patch.