Ticket #588, when a monster's bullwhip grabs your wielded in-use leash.
This is similar in principle to C343-262, only the leash being removed
from your inventory was due to a different cause.  There were several
reasonable ways I could have gone with the fix:  the leash could have
been removed from both you and the pet, for example, or the whip could
have just slipped off the leash like it does an iron ball or a cursed
wielded weapon.  I chose to go with a third option:  the whip becomes
entangled in the leash and is pulled from the monster's grasp.

We should probably also test other ways items can be removed from
the player's inventory, e.g., by a nymph.