Ticket #588, apply Dev Team fix, nethack.rankin, Sun Feb 18 03:24:58 2007 +0000
This fix handles the slippery fingers case.  Instead of specifically
checking for two individual criteria that would prevent dropping the
wielded item, it instead just calls canletgo(), which already had a
check for the leash case.  With the fix, the leash slips from your
hands and is no longer wielded, but it's still tied to you and so
remains in inventory.

I did change the call to canletgo() slightly, so that it prints a
message telling you the leash is tied to you.  Otherwise, the "slips
from your hands" message seemed to imply that the item was being
dropped, when it fact it wasn't.  The "tied" message clarifies this.