Ticket 476, C343-206, Apply Dev Team fix, nethack.rankin Tue Aug 22 03:31:32 2006 +0000
When the player received an artifact as a divine gift while blind or
hallucinating, it was unconditionally placed on the discoveries list,
even though the player did not see it and thus dknown was never set on
the object itself.  This resulted in an inconsistency between the list
of discoveries and which items would be shown as type-identified in
other contexts (such as in inventory).