Ticket #477, failing to untrap while standing on stairs should not create web.
When failing to #untrap a monster from a web while standing on stairs,
don't create web on the stairs.  No other trap, to my knowledge, is
allowed to exist on stairs, and web is not generated on stairs in any
other fashion, so having the web on stairs this way feels
inconsistent.

It also was causing strange bugs.  Among other things, being trapped
in the web didn't stop you from going up or down the stairs, which
would always immediately unstick you.  Also, it's a glyph conflict and
makes it hard to navigate to the stairs using travel.  Additionally,
the highlight option for stairs was not designed to handle the case
where there is also a trap there and failed to highlight the stairs,
which makes them really hard to find, if you happen to forget that
they got webbed.  I've had this happen to me on NAO; spent hundreds of
turns searching the level and ended up having to dig down and go back
up to find the down stairs (although watching the ttyrec -- or if this
had happened in NH4 the replay -- would also have worked).

So now if you are standing on stairs when you fail to #untrap a
creature from a web, the web will not spread to your tile.