Ticket #477, failed untrap should not place webbing on an already-trapped tile.
Thanks to b_jonas for commenting on this case.

If the player is standing on a trap and fails to #untrap a web, don't
add web to the player's tile in that case.  Yes, it's a loophole, but
unless we want to refactor traps so that tiles can have more than one
each, this is better than the alternative.