Apply Dev Team fix for C343-200, nethack.rankin, Sun Aug 6 05:13:22 2006 +0000
fix #H166 - trapdoors on quest levels

A player had reported that it was possible to fall from above the
quest locate level to below it.  This patch restricts falling or
uncontrolled levelporting past quest locate when it hasn't been
visited yet.  This improves the odds of proper sequencing for the
quest messages and avoids messing them up due to circumstances beyond
the player's control.

Once the locate level has been reached or passed, it is no longer a
barrier to falling or random teleport (because the relevant messages
occur only on first visit).

Controlled teleport is still allowed to bypass the quest locate level
even when it hasn't been visited; in that case quest messages given
later may be odd; hopefully, players new enough and unspoiled enough
to actually need the information that the quest messages provide are
unlikely to controlled-levelport past quest locate.