Apply Dev Team fix for what I've labeled UNL343-030, nethack.rankin, Sun Nov 5 07:20:20 2006 +0000
This makes new objects shop-owned if they were created from a
shop-owned horn of plenty.

The dev team's patch also applied when tipping a horn of plenty, but
NH4 currently does not have tipping, and even if we did implement it
for containers we might not bother with it for horns.  So I didn't
apply the tipping-related portion.

I also didn't apply the iflags.suppress_price aspect of the fix,
because upon a quick review of all the places where the dev team's
code does anything with that flag, it appears that NH4 has no
corresponding anything.

The dev team's commit log entry suggests that wishes from shop-owned
wands should also create shop-owned items, but this commit does not
attempt to implement that.