Apply Dev Team refinement for fix for UNL-040, nethack.rankin, Thu Dec 28 02:09:54 2006 +0000
The fix that made statue traps always produce an awake, hostile
monster, as originally applied, could result in a hostile co-aligned
unicorn, which the player would then be penalized for killing.

This knock-on fix simply prevents statue traps from being generated
with a statue of a co-aligned unicorn.  That probably means you can
still get a hostile co-aligned unicorn from a statue trap on a bones
level, not sure if we care -- bones levels are supposed to be able to
have things that don't occur otherwise, so maybe it's fine.  The other
possibility would be to do as the original committer suggests (but
does not implement) and have set_malign set a flag that overrides the
luck penalty.  But the Dev Team's fix is simpler, and I opted to
simply apply that.

Extra Bonus Fix:  statue traps no longer count double toward
extinction.  The code that creates the trap generates a temporary
monster, from which certain things are copied such as gender and
inventory, but it then discards the temporary monster.  When the trap
triggers and the statue comes to life, the new monster that gets
created also counts.  The fix makes the temporary monster not count
toward extinction.  This is similar to UNL-013 but not identical.

The interaction between this fix and multiple RNGs is interesting.  I
believe statue traps are only generated at level generation time, and
thus I don't think player actions could cause a desync (except on
rng_main, which is not expected to sync), unless the players are of
different alignment.  But in case the players might be of different
alignment, or in case I'm missing something, I went ahead and took the
precaution of having the statue trap only use the provided (probably
level-specific) RNG once.  If it has to reroll the species, it reverts
to rng_main.  This can potentially give you one monster with species
chosen by the main rng, in the form of a statue trap, to replace a
coaligned unicorn that the level gen rng would have given you for the
trap.  Since triggering statue traps is at least in principle
avoidable, this seems like the lesser evil.