Clean up "old style" function definition by adding void for arguments.
I had inadvertently written an empty argument list in the Perl style,
as simply (), when in C that really should be (void).  The compiler
was complaining about it being "old style".  It seems more modern to
me, and also I'd think it would be right at home in the NetHack
codebase if it's old-style, but anyway this fixes it up.