Apply one little part of a Dev Team fix, nethack.rankin, Sat Jan 27 04:33:16 2007 +0000
This was originally commited as part of the fix for C343-245.  The
main aspect of that bug had already been fixed in NH4, see ticket 584.
The Dev Team's fix to dokick() is somewhat more complicated, but the
fix we already have is good enough there -- it effectively prevents
the problem code from actually being reached if the coordinates are
off the edge of the map.

However, there was this one other little related bit, in
feel_location(), which I have applied here.  It causes the function to
return early if the location is off the map.

While investigating this fix I also noticed something about the
drawbridge functions, but I have not addressed that here.