Attacking a floating eye in melee is now possible, but you sustain damage.
The floating eye's passive line-of-sight-based telepathic abilities
cause you to suffer hp damage for it (if the eye survives),
proportional to the damage you inflicted and also proportional to your
own experience level.  I believe this is strictly less likely to end
your game in one hit than the 3.4.3 behavior of paralyzing you for
basically forever, but if I have the formula even close to right it
should be enough damage to make players think twice about melee as a
strategy against eyes.

And it's simply *got* to be more interesting than behavior that 4.3
inherited from Acehack.

As before, reflection, blindness, or anything that blocks line of
sight will protect you.  However, free action will not protect you,
since there is no longer any flavor justification for that, as the
attack isn't paralyzing in nature.  (It's more like reverse empathy.)

Also, this commit cleans up a bit more of the aftermath from merging
the multi-rng changes from upstream.  (For example, one of my changes
was creating an object and to do so was calling a function that now
takes a specific rng as an additional argument, so I needed to amend
my change to pass it one.)