Implement the new reverse-empathy-based floating eye behavior for mhitm.
This is the monster-vs-monster case (including pet-vs-hostile).  The
formula is a bit different from player-vs-monster, because monster
hitpoints are on a different scale from player hitpoints.  The formula
I ended up settling on is 1dn, where n is two thirds of the
aggressor's max hitpoints.  Thus, a fully-healed monster cannot die in
one hit by attacking a floating eye, but death in two hits is
possible, regardless of monster level.