Generate small numbers of certain otherwise rare objects on early levels.
The items in question are things that don't need to be common, but
giving the player a small quantity of them on the early dungeon levels
makes them more accessible and thus more relevant without making them
so common as to be unbalancing.

The formula I've adopted is, each level in the main dungeon from depth
2 through depth 6, if it is a standard (non-special) level, has a 75%
chance to get one of these items, in addition to whatever was
generated at random.  (Early Delphi takes away one of those 75%
chances, but early Delphi is a boon of its own, so that does not seem
like a problem to me.)  If the 75% check passes, the object created
can be a scroll of scare monster (50% chance) or a sprig of wolfsbane,
eucalyptus leaf, or wand of enlightenment (1/6 chance each).

The scroll of scare monster is intended to help offset the NetHack
Fourk Elbereth nerf.  Wolfsbane of course is intended for dealing with
early lycanthropy.  The eucalyptus is unlikely to be particularly
useful in the very early game (unless it happens to be blessed), but
it could come in handy later.

The wand of enlightenment is the most questionable inclusion.  It is
arguable that -- when generated this way -- it should have fewer
charges than would ordinarily be expected (perhaps 3-6 charges).
We'll see how that looks after a bit of play testing I guess.