Address some special level balance issues in the Gnomish Mines.
Minetown is now (barring bones) guaranteed to contain a shop that will
give prices for scrolls.  (In College Town, the book store was already
guaranteed.  In some of the towns, the general store was already
guaranteed.  In the remaining ones, the general store has been changed
to guaranteed.)

Bazaar Town has been jazzed up a bit.  It now always features two
nymphs and some winter wolves and can potentially contain an armor
shop, a weapons shop, a jewelry shop, or even a scroll shop.

Some of the other Minetown versions are still really too similar to
one another.  Further changes may be coming.

Mine's End has had its wimpiest, most boring version overhauled:  The
Mimic of the Mines now contains significantly more loot (including a
touchstone, a tinning kit, and a 50% chance of a horn of plenty) and
has a 75% chance of a polytrap.