Enhance the differences between various versions of Minetown.
Several of the versions of Minetown that previously felt rather
similar to one another have been adjusted in different directions.

Also, I changed the single 75% chance polytrap in the Mimic of the
Mines to two independent 50% chances.  This still means the overall
chance of at least one polytrap on the level is 75%, but now there is
an overall 25% chance of there being two of the silly things.