The statue of Perseus now guarantees the shield and sack (and scimitar).
As a boon for players who choose to skip Sokoban, the statue of
Perseus is now guaranteed to always contain the sack and the shield of
reflection, so it is never necessary to take on the Castle with no
container and no reflection.  I went ahead and guaranteed the scimitar
as well, because scimitars get short shrift enough in this game.

The levitation boots have the same probabilities as before.  There are
something like two dozen other ways to get past the water, and not
having a single guaranteed every-time solution encourages creativity,
which is a good thing.