Add some items to some of the quests.
The Arch quest gets a couple of boomerangs, a grappling hook, and a tinning kit.
Additional grappling hooks are pointless until the infrastructure work is done to
make them do what we want, which requires at least the levels branch and probably
additional 4.4 code work.  So that will have to wait.

The Barbarian quest gets a guaranteed long sword.  This is mainly so that lawful
Barbarians (pending in a future commit) will be able to #dip for Excalibur.  In
most cases they'll do so well before the quest, but this is a fallback.

The Knight quest gets two lances, a saddle, and a spellbook of magic missile.
The usefulness of these items should be clear.

The Samurai quest gets some Ya.  I'm not sure if each of these will turn into
a stack or not.  If so, most of them will need to be removed, because it'll be
way too many.  This number should be about right if they don't become stacks.
I'll have to test that.  Oh, I also threw in a saddle.