Pass level to accessible() to avoid segfault in goodpos().
If we're going to call accessible() in places where level isn't
defined, it should't be relying on that variable, then.  Make it not
do that.

This was causing a segfault in NH4 *almost* 100% of the time on new
game creation.  I'm guessing something about the way levels are
handled differently in NH4 vs vanilla explains why the Dev Team got
away with this.

I also corrected levtyp to schar, to match the typ that is assigned to
it.  I don't happen to know how an schar differs in principle from an
int nor how and under what circumstances it might differ in any
tangible way, nor why the compiler didn't catch this (nor what the
point is of having such specifically-typed variables in the language
if it does NOT mean the compiler will catch things like this), but I'm
pretty sure I don't want to find out two and a half years from now
when somebody tries to compile on some esoteric platform and it breaks
horrifically.