Starting inventory adjustments: Archeologist
The chance of an oil lamp is improved to 50%, and otherwise you are
compensated with a boomerang in lieu of the lamp.

The redundant percentage-chance tin opener is removed, since Arcs
always get a tin opener now (since a couple of commits ago).

The low-percentage chance at a magic marker is removed, because it
doesn't really play into how the Arc early game is structured, and
because we don't want to encourage start scumming.