Starting inventory adjustments: Barbarian
The chance for Barbarians to start with a battle axe (and short sword)
is now 75%.  (The other 25% of the time they get the two-handed sword
and a regular axe.)

The percentage chance of an oil lamp is removed.  It seems
unnecessary, at least for now.  I may reconsider that if I ever
rebalance the quantity of candles in the Mines.