Ticket #594, part 1: split stack when eucalyptus is unblessed due to applying.
Right now, this is a relatively obscure bug, but if we end up giving
healers a starting stack of blessed eucalyptus as Class Overhaul
proposes, this would become a big deal.  Let's just do an end run
around that and fix the bug pre-emptively.

Doing so brought up the fact that splitobj() did not correctly handle
the case of a stack held in player inventory.  (Specifically, it did
not assign a new inventory letter to the new object.)  So I fixed
that as well -- it will also be needed for other related fixes, e.g.,
the dulling case also mentioned in the ticket.