C343-310, when applying an unpaid stack of potions of oil, split the stack.
This is really just another case of Ticket #594, stacks sometimes
aren't split when they should be.  It's obscure, because it's not very
common to have more than one potion of oil at a time, much less more
than one unpaid potion of oil at a time, and it's even more obscure in
NH4 where potions of oil come pre-identified and candles are common,
removing the really compelling reasons to apply potions of oil.  Still
might as well fix it while I'm looking at this stack-splitting issue.