Starting inventory adjustments:  Tourist
The number of starting darts is improved from 3.4.3.  (It was briefly
even /more/ improved in NH4, due to a typo, but that would have been
corrected the next time I pull down upstream changes.  Actually, this
is probably going to be a merge conflict, but I'll deal with it.)

The odds of a tinopener, leash, or towel are improved; and if you
don't get them, instead of a low-percentage chance at a marker, you
are guaranteed an oil lamp.  Being guaranteed one of tinopener, leash,
towel, or lamp may not seem like good compensation for even a small
chance at a marker, but we do not wish to encourage start scumming,
and the increased number of darts should help -- especially since
Fourk has slightly improved dart damage.  And you'll want those darts,
since Tourists can no longer multi-fire daggers, due to the recent
skill cap adjustments.  In fact, I don't know if the number of darts
I've enacted here will really be enough.  That'll need play testing.