Starting inventory adjustments:  Valkyrie
The starting spellbook has a 1/3 chance to be magic missile, instead
of force bolt.  Note that it is not clear whether this is better for a
Valkyrie.  It will scale better into the middle game, but it will be
more difficult to get off the ground in the early game, in terms of
failure rate.

In any case, there'll be a spellbook of cone of cold guaranteed
on the quest.

The percentage chance of a starting lamp is removed.