Wands of wishing should not be rechargeable.
This commit makes wands of wishing not rechargeable.  You get
the number of charges the wand is generated with, plus the wrest.

I intend to compensate for this by making other wish sources
very slightly more common, but I don't know if I will get that
done before the public announcement on the first.  We'll see.