Make rocks more common, and update the changelog.
This places rocks on the floor of rooms sometimes, partly for flavor
but mostly as backup ammo for sling users.  Sling is in other ways
worse than most ranged weapons, but at least you shouldn't have to
worry now about running out of rocks.

I also updated the changelog to document the first round of role
balance changes, the wand of wishing change, and the rocks.