Make monkeys more interesting.
This improves the chances for a monkey to steal something from you and
also gives them a one-time burst of speed when they do, to help them
get away.

I tested that it's still relatively easy to chase them down and kill
them, because the NetHack monster AI is not very good at fleeing.
(Compare to Brogue, where early monkeys can be a real nightmare,
mainly because they are so good at fleeing it is literally completely
impossible to ever catch them if you don't happen to have a ranged
attack; even blinking is usually not enough.  I have successfully
avoided duplicating that nightmare here -- they can be caught.)