Ticket #834, Apply Dev Team fix, C343-139, nethack.rankin, Tue Dec 6 03:19:33 2005 +0000
One of the rgrn regulars had reported that playing mastermind with the
castle drawbridge while blind yields a sequence of messages about
tumblers and gears that you hear when partly correct, but when you
play the complete passtune and the bridge opens, there was no message
if you were blind.

The fix ignores the question of whether you were using the passtune or
not and simply always prints a message about hearing chains and gears
any time you can hear but cannot see the bridge open or close.

Added bonus:  when the drawbridge is destroyed, scatter some debris
(in the form of chains, because NetHack already has such an object).

This applied fairly straightforwardly.  Apparently, apart from moving
the drawbridge code into dbridge.c, we haven't otherwise changed it
that much.

I left the quantity and position of the chains up to the main rng on
the grounds that they are unlikely to have much real significance,
the chain debris being primarily cosmetic, IMO.