Implement scrolls of wishing.
Scrolls of wishing are unwishable and cost too much ink to ever be
written with markers.  When you read a non-cursed one, if your luck
is good enough (same rules as for wand of wishing), you get one wish.

This commit does not actually put any scrolls of wishing in the game.
It just provides the capability to do so.