Rings dropped into sinks are not destroyed, only buried.
To reclaim the (often now identified) rings, the player simply has to
destroy the sink, dry up or otherwise destroy the resulting fountain,
and dig a pit.

This adds a way to identify rings early, if there is a sink, at the
cost of the sink.  Sinks being a limited resource, the question will
be how long to wait, to collect more rings, to ID with the same sink,
given that eventually you're going to have other ways to identify
things anyway, and the rings might be useful now, if only you knew
what they were.