Attempt to make quasits somewhat more interesting.
Quasits are now slightly faster, generate slightly less often but in
small groups, and their dexterity draining attack is buffed a little.

I have my doubts how effective this will be, but I want to try
something, because quasits are in danger of being declare Boring and
removed from the game otherwise, and then I'd have to figure out what
to replace them with in the knight quest.

The other possibility I have considered for them is to buff AD_DRDX
attacks in general, but I'm not sure exactly what to do to them.