Apply Dev Team fix, C343-296, nethack.rankin, Thu May 10 03:03:46 2007 +0000
Sunsword, since it is supposed to have a bonus against undead,
should be able to damage shades, but in 3.4.3, it was having its
damage against them reduced to 1 point, because it wasn't made
of silver.  The fix allows the bonus-vs-undead attribute of an
artifact weapon to bypass the silver requirement against shades.

This is the Dev Team's second fix for this bug.  The first one
was backed out because it inadvertently allowed other artifact
weapons (e.g., Fire Brand) to damage Shades as well (because
they did not have the specific resistance flag for it), which
was not intended.  So that fix was backed out, and this one was
implemented in its place.

Applied very cleanly to NH4, almost as if we hadn't done anything
much to change the code for either Sunsword or shades.