Apply remaining portion of Dev Team fix, C343-299, nethack.rankin, Thu May 17 06:30:18 2007 +0000
It was noted in rgrn that when conflict caused a tame engulfer to
swallow the hero, and the hero hit it to get out, a "program in
disorder" message could be produced, due to "placing defunct monster
onto map".

As explained in the Dev Team's commit log, the problem was caused by
hmon_hitmon(), which subtracted damage then did some other stuff
including calling abuse_pet() and monflee() if the monster was tame,
before calling killed().  So monflee() called release_hero(), which
called expels(), placing the hero and the engulfer back on the map.
In the case that the damage had been fatal, this triggered the
reported warning.

The most critical part of the fix (returning early from monflee()
if the monster is dead already) had already been done for us by
Jared Minch in e568e1b8, but that still left the tameness reduction
handling portion of the fix unapplied, which could be relevant in
case of revival.  So this commit applies that part of the fix.