Apply Dev Team fix, C343-307, nethack.rankin, Tue May 29 03:52:02 2007 +0000
One of the beta testers had noticed that a unicorn horn would
restore lost attributes even if they were supposedly fixed at
their values by a worn ring of sustain ability.

The fix makes unicorn horns not do that but explicitly allows the
spell and potion of restore ability, and prayer, to override fixed
ability (and, in the case of major-trouble prayer, try to uncurse the
ring).  The thinking here is that currently players never use the
spell or potion as intended, because everyone just always carries a
unicorn horn.  With this fix, players who like to wear a ring of
sustain ability may find some use for the restore spell or potion.

Personally, I'm not sure how effective this will actually be at
getting people to use restore ability, but it doesn't hurt anything.