Apply Dev Team fix, C343-312, nethack.rankin, Sat Jun 9 02:18:44 2007 +0000
Monster polearm use is handled by the mthrowu code, which wasn't
checking to see if another weapon was welded to the monster's hand.
So the monster would try to wield the polearm, fail because of the
welded other weapon, and then proceed to attack with the polearm
anyway.  The fix introduces a specific check:  if the monster has
a weapon welded to is hand currently, it can't use any other object
as a polearm.

Applying this was pretty straightforward.  The trickiest thing was
tracking down that W_WEP translates to W_MASK(os_wep).

One caveat about my application:  to determine whether the monster's
current weapon is welded, the Dev Team's fix calls mwelded(), appears
not to exist in NH4 and in turn calls will_weld(), which does exist in
NH4 but can't be called from here because of where it is defined.  Our
definition of will_weld() differs from the Dev Team's in terms of
whether ammunition will weld or not; for consistency, I copied what
our will_weld() does, rather than the Dev Team's.  But that will
become inconsistent if our defition of will_weld changes.  From a
code maintenance standpoint, it might be good to move the definition
of will_weld from wield.c into a header that weapon.c can use; but
this commit does not do that.