Don't crash if a kick kills a player and monster simultaneously
For example, you get splashed with acid as the result of a kick that
killed a spotted jelly, and die in the process.

This rearrangement means that the monster will survive, undamaged,
in bones files (which is better than the monster surviving on
negative HP in bones files). This is not ideal, but breaks a cycle
of "player has to be alive to kill monster, monster has to be alive
to kill player".

Found by the testbench.