Add more spears in monster inventory.
When a monster gets a spear in its starting inventory,
it now actually gets rne(2) spears.

Also, the relative lack of stout spears has been rectified,
by occasionally giving some to a dwarf.

This is intended to ensure that players who wish to use
spears will be able to find some and have a reasonable
chance to be able to find positively enchanted ones, even.