Fix some monster-killed-twice crash bugs
Two bugs here. One is that killing a monster via monster spell wasn't
aborting the attack sequence, meaning it could die again. The other is
that the touch of Death, when used on a monster, was returning the
wrong return value in the case that the instadeath happened and wasn't
lifesaved from.

Both of these were found by the testbench.