Apply Dev Team fix, C343-323, nethack.rankin, Thu Jul 19 07:02:46 2007 +0000
As noted in rgrn, objects created as death drops when killing a
monster over water were not suffering the "floor effects" of the
water.

This commit only addresses the issue for random death drops (as
opposed to monster-specific ones).

Added Bonus Simplification:  due to a previous fix, the code that
checks for what objects a small monster can death drop no longer needs
to special-case leashes or food rations, as they fall under the
corrected weight cap naturally now.  This removes a line of code.