Ticket #116 bit:  Golems don't have skin (or other organs).
Golems (other than flesh golems) don't have organs.  They have
whatever substance they are made of, in the shape of a body, which
gives them the body parts that are defined by shape and position
(hands, legs, etc.) but not organs (like skin and stomach).  Eyes are
arguable (dolls, for example, are often said to have eyes), but I
have chosen to err on the side of golems not having eyes.