Add three Soko4 levels and disable the hardest Soko1 level.
This commit adds three more Soko4 levels ("Open at the Top", with
a bag of holding; "Collecting Marbles" with an amulet of reflection;
and "Jumpin' All Around" with a ring of teleport control).

It also disables the hardest Soko1 level, "Through the Cracks", which
upon closer inspection seems really too hard for the first level of
Sokoban.  Not sure what I was thinking there.  I mean, it's harder
than any of the Soko2 levels and half the Soko3 ones.  I like its
design, because it's /different/ from the others, but it's too
difficult for Soko1 and not geographically large enough for the later
levels of the branch.  So it's out, for now, pending redesign.