Add a digit to the version number, making the leading "4." implicit.
I want to be able to do more than one release based on the same
version of upstream NetHack 4, without ending up with completely
unrelated version numbers.  For example, if I do two releases based on
4.3.0, I don't want to end up numbering one of them 4.3.1, because
then when upstream 4.3.1 comes out, that'll be horribly confusing.

So I added a digit.  To avoid messing with the internals of how
version numbers work very much (e.g, the data structure), the
overarching leading "4." part is implicit.  So for the moment our
VERSION_MAJOR is 3 and our VERSION_MINOR is 0, meaning this is a
version based on NetHack 4 version 4.3.0.  This gives me the "third"
digit in the version data structure to use as a release counter.

The main menu and the #version command both prepend the 4., so the
version number the user sees will be 4.3.0.1 for now.  Internally,
the game sees it as version 3.0.1, but the leading 4 is implied.