Caveman Revamp, Round One:  Big Stick, Mammoth
The Caveman quest artifact is now Big Stick, an artifact club.

Mastodons have been replaced with mammoths.  In addition to improving
(slightly) the flavor diversity of the q monster class, this is also
needed for the Caveman revamp, because we plan to put mammoths in the
caveman quest when we revamp it.