Add a scroll of wishing to the Wizard's Tower, top level, inside.
When discussing the changes with someone, I misremembered where I'd
put the two scrolls and told him Vlad's Tower and the Wizard's Tower.
In fact, the second scroll was in O-Town.  But I'd been trying to
decide where to put a third one anyway, and the Wizard's Tower seems
as good a choice as any.  So put one there.  Now there are three.
(I'm not sure what the odds are that any of them will become a stack.
I think most scrolls on the floor are not stacks, typically, so
probably this will not be a major problem.)