Allow travel to displace pets
Also improves the message for when an action is aborted due to a
monster blocking the square (and a uim that aborts when it sees a
monster). This can be produced in a wide range of contexts;
"There's a monster there", the old message, doesn't work for
multi-turn actions when the monster wasn't visible at the start,
so try a new message "There's a monster in the way".