Valkyrie starting inventory adjustment:  add war hammer.
This gives Valkyries a starting war hammer, mainly for flavor.
Additionally, if the starting spellbook is magic missile (which
is harder for them to cast), then the +3 spear is silver, which
is a strict upgrade.

Valkyries will need further adjustments -- although, differentiating
them from Wizards may be better done by changing Wizards.