Make certain types of monsters drop corpses reliably for the first hundred kills.
The main motivation here was to guarantee lizard corpses as a stoning
remedy (assuming lizards are generated).  100 ought to be plenty to
last all game, so after that it uses the same formula as everything
else.

Since lizard corpses share non-spoiling flavor with acid blob and
lichen corpses, I added those too, and while I was at it I made it
apply to all F and all P, for flavor.

Then I added T to the list, so their corpses can revive.  What fun,
after all, would trolls be if they didn't drop corpses?  That's the
main thing that's special about them.

Also, I took 18/** out of a comment because it causes a warning,
and it wasn't really necessary.