Move the two "too-hard" Sokoban levels to later in Sokoban.
Earlier, I commented out the infamous "Soko2a" (nethackwiki, which
Steelypips calls 2 B), which is clearly too hard for the second
Sokoban level the player encounters.  Then at some point I decided
that Through the Cracks is too hard to be the first Sokoban level,
and I commented that one out too.

This commit restores brings them back, as possibilities for the third
level in Sokoban.