Clean up the food-appraisal-for-tins fix for NH4.
In NH4, food appraisal just makes edibility_prompts() more precise; it
isn't absolutely required.  Clean up the fix to account for that, both
in terms of the code and the comments.

I'm not sure this has any actual _effect_ at the moment, because
edibility_prompts() does require appraisal in order to warn about
tins.  (This hasn't changed.  The code for handling tins was there in
edibility_prompts all along; it just wasn't doing anything because
edibility_prompts() wasn't being called when you ate a tin.)

But the code should be more clear now if nothing else.