Avoid cross-aligned unicorn statue traps, and use msethostility() to set hostility.
ais523 suggested avoiding all unicorns as statue-trap monsters, not
just co-aligned ones, on the grounds that the trap may end up in a
bones level.  That probably wouldn't be worth much extra complication
in the code (because, after all, bones levels are /expected/ to be
potentially more trouble than regular levels), but in this case it
actually ends up simplifying the code, so my argument against it
pretty much falls apart.  So this commit makes that change.

Using the msethostility() function is an obvious win, now that I know
it exists.  Again, thanks to ais523 for pointing this out.