Get upward-thrown egg all over your face (or visor) even if it's not a cockatrice egg.
Previously, throwing egg up in the air would only give the "You've got
it all over your face" message if it was a petrifying egg and you were
vulnerable to petrification.  The fix to make a visored helmet protect
you changes "face" to "visor" in that case but still only gives the
message if the egg would have petrified you.

This commit puts the petrification check inside the other checks
(previously, it was wrapped around them), so now you get the message
about the egg being all over your visor or face first, and then (in
the latter case) the game checks to see if this petrifies you.

(I've said "face" several times here, but it actually uses the body
parts mechanism, so "face" can be something else if you are
polymorphed; that aspect of the logic hasn't changed at all.)